CU-85zrx08ta: Integrate Eveara Distribution Service addtl. work

+++ newm-server/src/main/kotlin/io/newm/server/features/distribution/model/GetUserSubscriptionResponse.kt
@@ -8,7 +8,7 @@ data class GetUserSubscriptionResponse(
     @SerialName("message")
     val message: String,
     @SerialName("total_records")
-    val totalRecords: Int,
+    val totalRecords: Long,
     @SerialName("success")
     val success: Boolean,
     @SerialName("data")
